% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_test.R
\name{c_test}
\alias{c_test}
\title{R frontend of the C function bar}
\usage{
c_test(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector x^2
}
\description{
A mask function to test if the dummy C code in src/bar.c
works.
}
\details{
\code{c_test(x)} = x^2.
}
\keyword{internal}
