% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{docu_test_function}
\alias{docu_test_function}
\title{Example function: roxygen2 documentation}
\usage{
docu_test_function(a, b)
}
\arguments{
\item{a}{A double}

\item{b}{A double}
}
\value{
A double a + b
}
\description{
This is the roxygen documentation of an example function.
}
