% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doxygen_tools.R
\name{DoxInit}
\alias{DoxInit}
\title{Prepares the R package structure for use with doxygen}
\usage{
DoxInit(rootFolder = ".")
}
\arguments{
\item{rootFolder}{The root of the R package}
}
\description{
Makes a configuration file in inst/doxygen
    and set a few options:
    \itemize{
       \item{EXTRACT_ALL = YES}
       \item{INPUT = src/}
       \item{OUTPUT_DIRECTORY = inst/doxygen/}
       \item{EXTRACT_PRIVATE = YES}
    }
}
\examples{
\dontrun{
DoxInit()
}
}
