% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doxygen_tools.R
\name{doxy_edit}
\alias{doxy_edit}
\title{Edits an existing Doxyfile}
\usage{
doxy_edit(options = c(), pkgFolder = ".",
  pathToDoxyfile = "inst/doc/doxygen/Doxyfile")
}
\arguments{
\item{options}{A named vector with new settings. The names represent
the tags.
A list of options can be found here:
\url{https://www.stack.nl/~dimitri/doxygen/manual/config.html}}

\item{pkgFolder}{A string with the path to any subfolder of the \R package root directory.  Default: \code{"."}.}

\item{pathToDoxyfile}{A string with the path to the \code{Doxyfile} relative to the package root folder indicated by \code{pkg}.  Default: \code{"inst/doc/doxygen/Doxyfile"}.}
}
\value{
\code{NULL}
}
\description{
Changes options in doxygen config files.
}
\examples{

\dontrun{
doxy_edit(options = c("EXTRACT_PRIVATE" = "YES"))
}

}
