% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doxygen_tools.R
\name{doxy_vignette}
\alias{doxy_vignette}
\title{Creates a doxygen vignette}
\usage{
doxy_vignette(pkg = ".", index = "html", viname, vientry)
}
\arguments{
\item{pkg}{A string with the path to any subfolder of the \R package root 
directory. Default: \code{"."}.}

<<<<<<< HEAD
\item{index}{A string with the path relative to \code{inst/doc} of the 
doxygen \code{index.html} file. Default: \code{doxygen/html} (see \strong{Note}).}
=======
\item{index}{A string with the path relative to \code{inst/doxygen} of the doxygen \code{index.html} file. Default: \code{html} (see \strong{Note}).}
>>>>>>> 90c921e4e5633f5958714269255d575e9a09ebfb

\item{viname}{A string giving the name of the \code{.Rmd} vignette file 
wrapping the documentation, as well as the name by which to retrieve the 
documentation using \code{vignette()}.  Default: \code{"pkgName-Doxygen.Rmd"}.}

\item{vientry}{A character string specifying the vignette Index Entry to use.
Default: "pkgName C++ library documentation".}
}
\value{
\code{NULL}
}
\description{
Creates an \R Markdown wrapper for the doxygen documentation, so that it 
can be viewed from within \R with a call to \code{vignette()}.
}
\details{
<<<<<<< HEAD
This function creates the file \code{vignettes/viname.Rmd} in the package 
root folder, containing the necessary meta-data for viewing the Doxygen HTML 
documentation from within \R with a call to \code{vignette()}.  When the vignette 
is built (e.g., with \code{R CMD build} or \code{devtools::build_vignettes()}), 
a file \code{inst/doc/viname.html} is created, and it is this file which is opened 
by the call to \code{vignette("viname")} after the package is installed. 
The contents of \code{inst/doc/viname.html} are simply a "redirect" to the Doxygen 
index file, \code{inst/doc/pathToIndex/index.html}.
}
\note{
The call to \code{vignette()} will *only* open HTML files stored in the 
\code{doc} subfolder of an installed package.  Therefore the Doxygen documentation 
referred to by \code{pathToIndex} must be stored in a subfolder of \code{inst/doc} 
for the call to \code{vignette()} post-installation to work.
=======
This function creates the file \code{vignettes/viname.Rmd} in the package root folder, containing the necessary meta-data for viewing the Doxygen HTML documentation from within \R with a call to \code{vignette()}.
}
\note{
The call to \code{vignette()} will *only* open HTML files stored in the \code{doc} subfolder of an installed package.  Therefore, a natural location for the doxygen documentation (doxydoc) is in \code{inst/doc/doxygen}.  However, the latest version of \pkg{devtools} incontrovertibly deletes \code{inst/doc} during the build/install process.  Due to the ubiquitous usage of \pkg{devtools} among \R package developers, the doxydoc is stored here in \code{inst/doxygen}, and during the build process, moved (or technically, copied and source added to \code{.Rbuildignore}) via a \code{vignettes/Makefile}.  Packages with their own such \code{Makefile} will not have it overwritten, and developers may view the default \code{Makefile} provided by \pkg{rdoxygen} with the call
\preformatted{
cat(readLines(system.file("sys", "Makefile",
                          package = "rdoxygen")), sep = "\n")
}
>>>>>>> 90c921e4e5633f5958714269255d575e9a09ebfb
}
