% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doxygen_tools.R
\name{ReplaceTag}
\alias{ReplaceTag}
\title{Replace a value for a given tag on file in memory}
\usage{
ReplaceTag(fileStrings, tag, newVal)
}
\arguments{
\item{fileStrings}{A vector with each string containing a line of the file}

\item{tag}{The tag to be searched for}

\item{newVal}{The new value for the tag}
}
\value{
The vector of strings with the new value
}
\description{
Scan the lines and change the value for the named tag if one line has this tag,
   add a line at the end if no line has this tag and return a warning if several lines
   matching the tag
}
\examples{
\dontrun{
fakeFileStrings <- c("Hello = world","SURE\\t= indeed","Hello = you")

expect_warning(ReplaceTag(fakeFileStrings,"Hello","me"))

newFake <- ReplaceTag(fakeFileStrings,"SURE","me")
expect_equal(length(newFake), length(fakeFileStrings))
expect_equal(length(grep("SURE",newFake)), 1)
expect_equal(length(grep("me",newFake)), 1)

newFake <- ReplaceTag(fakeFileStrings,"Bouh","frightened?")
expect_equal(length(newFake), length(fakeFileStrings)+1)
expect_equal(length(grep("Bouh",newFake)), 1)
expect_equal(length(grep("frightened?",newFake)), 1)
}
}
