% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doxygen_tools.R
\name{doxy}
\alias{doxy}
\title{Calls doxygen for an \R package}
\usage{
doxy(pkgFolder = ".", pathToDoxyfile = "inst/doc/doxygen/Doxyfile")
}
\arguments{
\item{pkgFolder}{A string with the path to any subfolder of the \R package root directory.  Default: \code{"."}.}

\item{doxyfile}{A string with the path to the \code{Doxyfile} relative to the package root folder indicated by \code{pkg}.  Default: \code{"inst/doc/doxygen/Doxyfile"}.}
}
\value{
\code{NULL}
}
\description{
Creates doxygen documentation based on the given \code{Doxyfile} configuration.  Creates this file with default values (using \code{doxy_init()}) if it doesn't exist.
}
\examples{
\dontrun{
  doxy()
}

}
