% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doxygen_tools.R
\name{doxy}
\alias{doxy}
\title{devtools document function when using doxygen}
\usage{
doxy(doxygen = file.exists("src"))
}
\arguments{
\item{doxygen}{A boolean: should doxygen be ran on documents in src?
the default is TRUE if a src folder exist and FALSE if not}
}
\value{
The value returned by devtools::document()
}
\description{
Overwrites devtools::document() to include the treatment of
   doxygen documentation in src/
}
\examples{
\dontrun{
doxy()
}
}
