% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doxygen_tools.R
\name{doxy_edit}
\alias{doxy_edit}
\title{Edits an existing Doxyfile}
\usage{
doxy_edit(pkg = ".", doxyfile = "inst/doxygen/Doxyfile",
  options = c())
}
\arguments{
\item{pkg}{A string with the path to any subfolder of the \R package root 
directory. Default: \code{"."}.}

\item{doxyfile}{A string with the path to the \code{Doxyfile} relative 
to the package root folder indicated by \code{pkg}.  
Default: \code{"inst/doxygen/Doxyfile"}.}

\item{options}{A named vector with new settings. The names represent the tags.
A list of options can be found here: \url{http://www.doxygen.nl/manual/config.html}.}
}
\value{
\code{NULL}
}
\description{
Changes options in doxygen config files.
}
