% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doxygen_tools.R
\name{doxy_init}
\alias{doxy_init}
\title{Prepares the R package structure for use with doxygen}
\usage{
doxy_init(rootFolder = ".")
}
\arguments{
\item{rootFolder}{A string with the path to the root directory of the R
package. Default: "."}
}
\description{
Creates a configuration file in inst/doxygen/ and sets a few options:
    \itemize{
       \item{EXTRACT_ALL = YES}
       \item{INPUT = src/}
       \item{OUTPUT_DIRECTORY = inst/doxygen/}
    }
}
\examples{

\dontrun{
doxy_init()
}

}
