% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doxygen_tools.R
\name{doxy_init}
\alias{doxy_init}
\title{Prepares the R package structure for use with doxygen}
\usage{
doxy_init(pkg = ".", doxyfile = "inst/doxygen/Doxyfile")
}
\arguments{
\item{pkg}{A string with the path to any subfolder of the \R package root directory.  Default: \code{"."}.}

\item{doxyfile}{A string with the path to the \code{Doxyfile} relative to the package root folder indicated by \code{pkg}.  Default: \code{"inst/doc/doxygen/Doxyfile"}.}
}
\value{
\code{NULL}.
}
\description{
Creates a Doxygen configuration file and sets a few options:
    \itemize{
       \item{\code{INPUT = src/ inst/include}}
       \item{\code{OUTPUT_DIRECTORY = inst/doc/doxygen/}}
       \item{\code{GENERATE_LATEX = NO}}
       \item{\code{HIDE_UNDOC_MEMBERS = YES}}
       \item{\code{USE_MATHJAX = YES}}
       \item{\code{PROJECT_NAME = name_of_R_package}}
    }
}
\details{
For git users: by default the Doxygen documentation is stored in \code{inst/doc/doxygen}, but this folder is often added to \code{.gitignore} (e.g., by \code{devtools::use_vignettes}).  To unignore some or all of the Doxygen documentation, you may wish to add the following lines to the package \code{.gitignore}:
\preformatted{
# ignore inst/doc, but allow its contents to be selectively unignored
!inst/doc
inst/doc/*

# unignore just the Doxyfile (need Doxygen to recompile the docs)
!inst/doc/doxygen/Doxyfile

# unignore all the Doxygen documentation
!inst/doc/doxygen
}
}
\examples{

\dontrun{
doxy_init()
}

}
