% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doxygen_tools.R
\name{doxy_init}
\alias{doxy_init}
\title{Prepares the R package structure for use with doxygen}
\usage{
doxy_init(pkg = ".", doxyfile = "inst/doc/doxygen/Doxyfile")
}
\arguments{
\item{pkg}{A string with the path to any subfolder of the \R package root directory.  Default: \code{"."}.}

\item{doxyfile}{A string with the path to the \code{Doxyfile} relative to the package root folder indicated by \code{pkg}.  Default: \code{"inst/doc/doxygen/Doxyfile"}.}
}
\value{
\code{NULL}.
}
\description{
Creates a Doxygen configuration file and sets a few options:
    \itemize{
       \item{\code{INPUT = src/ inst/include}}
       \item{\code{OUTPUT_DIRECTORY = inst/doc/doxygen/}}
       \item{\code{GENERATE_LATEX = NO}}
       \item{\code{PROJECT_NAME = name_of_R_package}}
    }
}
\examples{

\dontrun{
doxy_init()
}

}
