% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doxygen_tools.R
\name{doxy_init}
\alias{doxy_init}
\title{Prepares the R package structure for use with doxygen}
\usage{
doxy_init(pkg = ".", doxyfile = "inst/doxygen/Doxyfile")
}
\arguments{
\item{pkg}{A string with the path to any subfolder of the \R package root 
directory. Default: \code{"."}.}

\item{doxyfile}{A string with the path to the \code{Doxyfile} relative 
to the package root folder indicated by \code{pkg}.  
Default: \code{"inst/doxygen/Doxyfile"}.}
}
\value{
\code{NULL}.
}
\description{
Creates a Doxygen configuration file and sets a few options:
    \itemize{
       \item{\code{INPUT = src/ inst/include}}
       \item{\code{OUTPUT_DIRECTORY = inst/doxygen/}}
       \item{\code{GENERATE_LATEX = NO}}
       \item{\code{HIDE_UNDOC_MEMBERS = YES}}
       \item{\code{USE_MATHJAX = YES}}
       \item{\code{PROJECT_NAME = name_of_R_package}}
    }
}
<<<<<<< HEAD
=======
\details{
While the package developer is free to change the \code{OUTPUT_DIRECTORY} to wherever they like, the default value above is suggested for compatibility with the \pkg{devtools} package authoring workflow.  See \code{\link{doxy_vignette}} for details.
}
>>>>>>> 90c921e4e5633f5958714269255d575e9a09ebfb
\examples{

\dontrun{
doxy_init()
}

}
