% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doxygen_tools.R
\name{replace_tag}
\alias{replace_tag}
\title{Updates and adds doxygen options in a line string vector}
\usage{
replace_tag(fileStrings, tag, newVal)
}
\arguments{
\item{fileStrings}{A vector with each string containing a line of the 
file}

\item{tag}{A string with the tag to be searched for}

\item{newVal}{A string with the new value for the tag}
}
\value{
The vector of strings with the new value
}
\description{
Scans the lines and changes the value for the named tag if one line has 
this tag, adds a line at the end if no line has this tag and returns a 
warning if several lines match the tag.
}
